Refactor: split proc macro host into submodules (#1747)
commit-id:6d8f33f9

---

**Stack**:
- #1749
- #1748
- #1745
- #1747 ⬅

⚠️ *Part of a stack created by [spr](https://github.com/ejoffe/spr). Do
not merge manually using the UI - doing so may have unexpected results.*
